ttkd_api/requirements.txt
Django==1.10
djangorestframework==3.4.7
PyInstaller==3.2
pylint==1.6.4
pylint-django==0.7.2
pylint-plugin-utils==0.2.4
django-filter==0.15.2
Pillow==3.4.2
djangorestframework-csv==2.0.0